+++ cli/cli/App.cs
@@ -21,6 +21,7 @@ using cli.Portal;
 using cli.Services;
 using cli.Services.Content;
 using cli.Services.HttpServer;
+using cli.TempCommands;
 using cli.TokenCommands;
 using cli.UnityCommands;
 using cli.Unreal;
@@ -38,11 +39,13 @@ using Serilog.Enrichers.Sensitive;
 using Serilog.Events;
 using Spectre.Console;
 using System.CommandLine;
+using System.CommandLine.Binding;
 using System.CommandLine.Builder;
 using System.CommandLine.Help;
 using System.CommandLine.Invocation;
 using System.CommandLine.IO;
 using System.CommandLine.Parsing;
+using System.Diagnostics;
 using System.Reflection;
 using Command = System.CommandLine.Command;
 
@@ -83,18 +86,10 @@ public class App
 	{
 		
 		var appCtx = provider.GetService<IAppContext>();
-		try
-		{
-			if (appCtx.ShouldUseLogFile)
-			{
-				File.Delete(appCtx.TempLogFilePath);
-			}
-		}
-		catch
-		{
-			// if we cannot delete the temp file, then the worst outcome is that we are appending new data to it.
-		}
-
+		var config = provider.GetService<ConfigService>();
+		var binding = provider.GetService<BindingContext>();
+		config.SetupBasePath(binding);
+		
 		// https://github.com/serilog/serilog/wiki/Configuration-Basics
 		configureLogger ??= config =>
 		{
@@ -115,9 +110,27 @@ public class App
 					.WriteTo.BeamAnsi("{Message:lj}{NewLine}{Exception}")
 					.MinimumLevel.ControlledBy(app.LogLevel)
 			);
-			if (appCtx.ShouldUseLogFile)
+			if (appCtx.ShouldUseLogFile && appCtx.TryGetTempLogFilePath(out var logPath))
 			{
-				baseConfig.WriteTo.File(appCtx.TempLogFilePath, LogEventLevel.Verbose);
+				var path = Path.GetDirectoryName(logPath);
+				if (Directory.Exists(path))
+				{
+					var existingLogFiles = Directory.GetFiles(path);
+					// this is magic number... I guess its a rough estimate of a number of commands per day?
+					const int MaxNumberOfLogFilesBeforeAutoClean = 250;
+					if (existingLogFiles.Length > MaxNumberOfLogFilesBeforeAutoClean)
+					{
+						// clean up everything older than a day
+						ClearTempLogFilesCommand.CleanLogs(TimeSpan.FromDays(1), existingLogFiles);
+					}
+				}
+				baseConfig.WriteTo.File(logPath, LogEventLevel.Verbose);
+			}
+
+			if (appCtx.ShouldEmitLogs)
+			{
+				baseConfig.WriteTo.Sink(new ReporterSink(provider))
+					.MinimumLevel.ControlledBy(app.LogLevel);
 			}
 
 			return baseConfig.CreateLogger();
@@ -196,6 +209,20 @@ public class App
 			_setLogger = (provider) =>
 			{
 				ConfigureLogging(this, provider, configureLogger);
+				
+				TaskLocalLog.Instance.globalLogger = Log.Logger;
+				Log.Logger = TaskLocalLog.Instance;
+			};
+		}
+		else
+		{
+			_setLogger = (provider) =>
+			{
+				var appCtx = provider.GetService<IAppContext>();
+				if (appCtx.ShouldEmitLogs)
+				{
+					TaskLocalLog.Instance.CreateContext(provider);
+				}
 			};
 		}
 
@@ -251,6 +278,7 @@ public class App
 			root.AddGlobalOption(UnmaskLogsOption.Instance);
 			root.AddGlobalOption(NoLogFileOption.Instance);
 			root.AddGlobalOption(DockerPathOption.Instance);
+			root.AddGlobalOption(EmitLogsOption.Instance);
 			root.AddGlobalOption(provider.GetRequiredService<ConfigDirOption>());
 			root.AddGlobalOption(provider.GetRequiredService<ShowRawOutput>());
 			root.AddGlobalOption(provider.GetRequiredService<ShowPrettyOutput>());
@@ -314,6 +342,10 @@ public class App
 		Commands.AddSubCommand<GetTokenViaRefreshCommand, GetTokenViaRefreshCommandArgs, TokenCommandGroup>();
 		Commands.AddSubCommand<GetTokenForGuestCommand, GetTokenForGuestCommandArgs, TokenCommandGroup>();
 
+		Commands.AddRootCommand<TempCommandGroup>();
+		Commands.AddSubCommand<TempClearCommandGroup, TempClearArgs, TempCommandGroup>();
+		Commands.AddSubCommandWithHandler<ClearTempLogFilesCommand, ClearTempLogFilesCommandArgs, TempClearCommandGroup>();
+		
 		Commands.AddRootCommand<PlayerCommand, PlayerCommandArgs>();
 		Commands.AddSubCommandWithHandler<AddPlayerToRealmCommand, AddPlayerToRealmCommandArgs, PlayerCommand>();
 		
@@ -435,7 +467,11 @@ public class App
 
 		CommandProvider = Commands.Build();
 
+		var sw = new Stopwatch();
+		sw.Start();
 		var _ = InstantiateAllCommands();
+		sw.Stop();
+		Log.Verbose("BUILD ALL COMMANDS TOOK : " + sw.ElapsedMilliseconds);
 
 		// sort the commands
 		var root = CommandProvider.GetService<RootCommand>();
@@ -639,6 +675,8 @@ public class App
 		
 		commandLineBuilder.AddMiddleware(async (ctx, next) =>
 		{
+			var sw = new Stopwatch();
+			sw.Start();
 			// create a scope for the execution of the command
 			var provider = CommandProvider.Fork(services =>
 			{
@@ -649,16 +687,21 @@ public class App
 				ConfigureServices(services);
 			});
 
+			Log.Verbose("command prep (make provider) took " + sw.ElapsedMilliseconds);
+
 			// update log information before dependency injection is sealed.
 			_setLogger(provider);
+
+			Log.Verbose("command prep (make logs) took " + sw.ElapsedMilliseconds);
 
 			// we can take advantage of a feature of the CLI tool to use their slightly jank DI system to inject our DI system. DI in DI.
 			ctx.BindingContext.AddService(_ => new AppServices { duck = provider });
 			var appContext = provider.GetRequiredService<IAppContext>();
 			appContext.Apply(ctx.BindingContext);
 
-			// Check if we need to forward this command
-			// we only forward if the executing version of the CLI is different from the one locally installed on the project.
+			Log.Verbose("command prep (app context) took " + sw.ElapsedMilliseconds);
+
+			// Check if we need to forward this command --- we only forward if the executing version of the CLI is different than the one locally installed on the project.
 			// As long as the versions are the same, running the local one or the global one changes nothing in behaviour.
 			var runningVersion = appContext.ExecutingVersion;
 			var localVersion = appContext.LocalProjectVersion;
@@ -674,6 +717,8 @@ public class App
 				}
 			}
 			
+			Log.Verbose("command prep (past proxy) took " + sw.ElapsedMilliseconds);
+			
 			try
 			{
 				var beamoSystem = provider.GetService<BeamoLocalSystem>();
@@ -688,8 +733,11 @@ public class App
 				throw;
 			}
 
+			Log.Verbose("command prep took " + sw.ElapsedMilliseconds);
 			await next(ctx);
 
+			sw.Stop();
+			Log.Verbose("command execution took " + sw.ElapsedMilliseconds);
 		}, MiddlewareOrder.Configuration);
 		commandLineBuilder.UseDefaults();
 		commandLineBuilder.UseSuggestDirective();
@@ -746,9 +794,17 @@ public class App
 
 			if (appContext.ShouldUseLogFile)
 			{
-				Log.CloseAndFlush();
-				File.AppendAllText(appContext.TempLogFilePath, ex.ToString());
-				Console.Error.WriteLine("\nLogs at\n  " + appContext.TempLogFilePath);
+				if (!appContext.TryGetTempLogFilePath(out var logFile))
+				{
+					Log.Warning("Could not write logs because no .beamable project exists to host the log file. Please re-run the command with '--logs v', or, create a .beamable folder using 'beam init' and then re-run the command.");
+				}
+				else
+				{
+					Log.CloseAndFlush();
+					
+					File.AppendAllText(logFile, ex.ToString());
+					Console.Error.WriteLine("\nLogs at\n  " + Path.GetFullPath(logFile));
+				}
 			}
 		});
 		return commandLineBuilder.Build();
@@ -929,7 +985,11 @@ public class App
 
 	public virtual Task<int> RunWithSingleString(string commandLine)
 	{
+		var sw = new Stopwatch();
+		sw.Start();
 		var prog = GetProgram();
+		sw.Stop();
+		Log.Verbose("prepared virtual program instance in " + sw.ElapsedMilliseconds);
 		return prog.InvokeAsync(commandLine);
 	}
 }