+++ client/Packages/com.beamable.server/Editor/Usam/Editor/CodeService.cs
@@ -931,7 +931,6 @@ namespace Beamable.Server.Editor.Usam
 				return;
 			}
 
-			Promise errorPromise = new Promise();
 			string errorMessage = string.Empty;
 
 			var args = new ProjectNewServiceArgs()
@@ -944,7 +943,6 @@ namespace Beamable.Server.Editor.Usam
 			var command = _cli.ProjectNewService(args).OnError((cb) =>
 			{
 				errorMessage = $"Error creating service: {serviceName}. Message=[{cb.data.message}] Stacktrace=[{cb.data.stackTrace}]";
-				errorPromise.CompleteSuccess();
 			});
 			await command.Run();
 
@@ -952,11 +950,10 @@ namespace Beamable.Server.Editor.Usam
 
 			var definition = ServiceDefinitions.FirstOrDefault(def => def.BeamoId == serviceName);
 
-			if (definition == null)
+			if (definition == null || !string.IsNullOrEmpty(errorMessage))
 			{
-				await errorPromise;
 				errorCallback?.Invoke();
-				throw new Exception(errorMessage);
+				throw new Exception(errorMessage ?? "no error message, but definition was null somehow");
 			}
 
 			if (assemblyReferences != null)